+++ test/log_tests.jl
@@ -6,7 +6,6 @@
     params = Dict{String, Any}(
         "num_nodes" => 5,
         "instance" => 1,
-        "optimizer" => "glpk",
         "tol_zero" => 1E-4,
         "tol_psd" => 1E-3,
         "eigen_cuts_full" => true,
@@ -14,10 +13,9 @@
         "lazycuts_logging" => true
     )
 
-    lom_optimizer = get_solver(params)
     data = LOpt.get_data(params)
     model_lopt  = LOpt.build_LOModel(data)
-    result_lopt = LOpt.optimize_LOModel!(model_lopt, optimizer = lom_optimizer)
+    result_lopt = LOpt.optimize_LOModel!(model_lopt, optimizer = glpk_optimizer)
 
     LOpt.visualize_solution(result_lopt, data, visualizing_tool = "tikz", plot_file_format = "tex")
     LOpt.visualize_solution(result_lopt, data, visualizing_tool = "tikz", plot_file_format = "tex", display_edge_weights=true)